init util.py and add playing/guessing methods
